% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcat.comp.R
\name{cdcat.comp}
\alias{cdcat.comp}
\title{Comparison of multiple cdcat objects}
\usage{
cdcat.comp(cdcat.obj.l, alpha, ...)
}
\arguments{
\item{cdcat.obj.l}{List of \code{cdcat} objects to be compared.}

\item{alpha}{N x K matrix with the attribute patterns to be compared to the \code{cdcat} results.}
}
\value{
\code{cdcat.comp} returns an object of class \code{cdcat.comp}.
}
\description{
This function compares differet \code{cdcat} objects in terms of classification accuracy (FIXED.LENGTH == TRUE) and/or CAT lenght FIXED.LENGTH == FALSE).
}
