% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR2step.R
\name{LR_2step}
\alias{LR_2step}
\title{Item-level model comparison using 2LR test}
\usage{
LR_2step(GDINA.obj, ...)
}
\arguments{
\item{GDINA.obj}{Calibrated item bank. An estimated model object of class GDINA}
}
\value{
\code{LR2step} returns an object of class \code{LR2step}.
}
\description{
This function evaluates whether the saturated G-DINA model can be replaced by reduced CDMs without significant loss in model data fit for each item using two-step likelihood ratio test (2LR). Sorrel, de la Torre, Abad, and Olea (2017) and Ma & de la Torre (2018) can be consulted for details.
}
\examples{
N <- 500
Q <- sim30GDINA$simQ
J <- nrow(Q)
gs <- data.frame(guess=rep(0.1,J),slip=rep(0.1,J))

sim <- simGDINA(N,Q,gs.parm = gs,model = "DINA")
resGDINA <- GDINA(dat = sim$dat, Q = sim$Q,
                 model = "GDINA",verbose = F)
LR2 <- LR_2step(GDINA.obj = resGDINA);round(LR2$LR2.p, 2)
}
\references{
Ma, W. & de la Torre, J. (2018). Category-level model selection for the sequential G-DINA model. \emph{Journal of Educational and Behavorial Statistic, 44}, 45-77.

Sorrel, M. A., de la Torre, J., Abad, F. J., & Olea, J. (2017). Two-step likelihood ratio test for item-level model comparison in cognitive diagnosis models. \emph{Methodology, 13}, 39-47.
}
