% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim180DINA.R, R/sim180GDINA.R
\docType{data}
\name{sim180DINA}
\alias{sim180DINA}
\title{Simulated data (180 items, DINA model)}
\format{A list with components:
\describe{
\item{\code{simdat}}{Simulated responses of 500 examinees}
\item{\code{simQ}}{Artificial Q-matrix}
\item{\code{simcatprob.parm}}{Artificial item parameters (probability of success for each latent group). All items are DINA items}
\item{\code{simalpha}}{Simulated attribute patterns of 500 examinees)}
}}
\usage{
sim180DINA

sim180DINA
}
\description{
Simulated data, Q-matrix and item parameters for a 180-item bank with 5 attributes.

Simulated data, Q-matrix and item parameters for a 180-item bank with 5 attributes.
}
\keyword{datasets}
