% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim180GDINA.R
\docType{data}
\name{sim180GDINA}
\alias{sim180GDINA}
\title{Simulated data (180 items, G-DINA model)}
\format{
A list with components:
\describe{
\item{\code{simdat}}{Simulated responses of 500 examinees}
\item{\code{simQ}}{Artificial Q-matrix}
\item{\code{simcatprob.parm}}{Artificial item parameters (probability of success for each latent group). All items are G-DINA items}
\item{\code{simalpha}}{Simulated attribute patterns of 500 examinees}
}
}
\usage{
sim180GDINA
}
\description{
Simulated data, Q-matrix and item parameters for a 180-item bank with 5 attributes.
}
\keyword{datasets}
