% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att.plot.R
\name{att.plot}
\alias{att.plot}
\title{Create plots for attribute mastery estimates}
\usage{
att.plot(cdcat.obj, i, k = NULL, ...)
}
\arguments{
\item{cdcat.obj}{An object of class \code{cdcat}}

\item{i}{Examinee to be plotted}

\item{k}{Attribute/s to be plotted. Default is NULL, which plotts all attributes}
}
\value{
\code{att.plot} creates a plot.
}
\description{
Create plots for attribute mastery estimates (\emph{X}: item position, \emph{Y}: mastery probability).
}
