% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcat.comp.R
\name{cdcat.comp}
\alias{cdcat.comp}
\title{Comparison of multiple cdcat objects}
\usage{
cdcat.comp(cdcat.obj.l, alpha, label = NULL)
}
\arguments{
\item{cdcat.obj.l}{List of \code{cdcat} objects to be compared}

\item{alpha}{N x K matrix with the attribute patterns to be compared to the \code{cdcat} results}

\item{label}{labels for the \code{cdcat} objects. If NULL (by default), the models are used as labels}
}
\value{
\code{cdcat.comp} returns an object of class \code{cdcat.comp}.
}
\description{
This function compares different \code{cdcat} objects in terms of classification accuracy (\code{FIXED.LENGTH == TRUE}) and/or CAT lenght (\code{FIXED.LENGTH == FALSE}).
}
