% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcat.summary.R
\name{cdcat.summary}
\alias{cdcat.summary}
\title{Summary information for a \code{cdcat} object}
\usage{
cdcat.summary(cdcat.obj, alpha)
}
\arguments{
\item{cdcat.obj}{An object of class \code{cdcat}}

\item{alpha}{N x K matrix with the attribute patterns to be compared to the \code{cdcat} results}
}
\value{
\code{cdcat.summary} returns an object of class \code{cdcat.summary}.
}
\description{
This function provides classification accuracy (\code{FIXED.LENGTH == TRUE)} and/or CAT length (\code{FIXED.LENGTH == FALSE}) results for cdcat object.
}
