% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.itembank.R.R
\name{gen.itembank}
\alias{gen.itembank}
\title{Item bank generation}
\usage{
gen.itembank(Q = NULL, minJ.K = NULL, complexity = NULL, IQ, VAR, model = NULL)
}
\arguments{
\item{Q}{Q-matrix}

\item{minJ.K}{Vector indicating the minimum number of items measuring each attribute}

\item{complexity}{Vector indicating the maximum number of attributes being measured by an item in each row of Q. At this moment maximum is 4}

\item{IQ}{Item discrimination (mean for the uniform distribution). IQ = \emph{P}(\strong{1}) - \emph{P}(\strong{0}) (Sorrel et al., 2017)}

\item{VAR}{Item discrimination (variance for the uniform distribution) (Kaplan et al., 2015)}

\item{model}{Vector indicating the model-item correspondence (0 = one-attribute item, 1 = DINA, 2 = DINO, 3 = A-CDM}
}
\value{
\code{gen.itembank} returns an object of class \code{gen.Item.Bank}.
}
\description{
This function can be used to generate an item bank. The user can provide a Q-matrix or create one defining the number of times each attribute should
be measured and the q-vector complexity (e.g., number of attributes that a q-vector can measure). Item parameters are sampled from
a uniform distribution with mean = \emph{IQ} and variance = \emph{VAR}.
}
\examples{
####################################
#        Example 1.                #
#     Q and model are provided     #
####################################

Q <- GDINA::sim30GDINA$simQ
model <- rep(1, each = nrow(Q))
IQ <- .70   # P(1), IQ = Low item quality in Kaplan, de la Torre & Barrada (2015)
VAR <- 0.10 # High variance in Kaplan et al. (2015)
bank <- gen.itembank(Q = Q, IQ = IQ, VAR = VAR, model = model)

####################################
#        Example 2.                #
#   Q and model are not provided   #
####################################

minJ.K  <- c(50, 50, 50)
complexity <- c(3, 3, 3)
IQ <- .70   # P(1), IQ = Low item quality in Kaplan, de la Torre & Barrada (2015)
VAR <- 0.10 # High variance in Kaplan et al. (2015)
bank <- gen.itembank(minJ.K = minJ.K, complexity = complexity, IQ = IQ, VAR = VAR)

}
\references{
Kaplan, M., de la Torre, J., & Barrada, J. R. (2015). New item selection methods for cognitive diagnosis computerized adaptive testing. \emph{Applied Psychological Measurement, 39}, 167-188.

Sorrel, M. A., Abad, F. J., Olea, J., de la Torre, J., & Barrada, J. R. (2017). Inferential item-fit evaluation in cognitive diagnosis modeling. \emph{Applied Psychological Measurement, 41}, 614-631.
}
