% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim180combination.R
\docType{data}
\name{sim180combination}
\alias{sim180combination}
\title{Simulated data (180 items, a combination of DINA, DINO, and A-CDM items)}
\format{
A list with components:
\describe{
\item{\code{simdat}}{Numeric array. Simulated responses of 250 examinees for two replicates}
\item{\code{simQ}}{Numeric matrix. Simulated Q-matrix}
\item{\code{simalpha}}{Numeric array. Simulated attribute patterns of 250 examinees for two replicates}
\item{\code{specifications}}{A list that contains all the specifications that were used in the \code{gen.itembank} function}
}
}
\usage{
sim180combination
}
\description{
Simulated data, Q-matrix and item parameters for a 180-item bank with 5 attributes. Data generated using the \code{gen.data} function.
}
\keyword{datasets}
