% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim180DINA.R
\docType{data}
\name{sim180DINA}
\alias{sim180DINA}
\title{Simulated data (180 items, DINA model)}
\format{
A list with components:
\describe{
\item{\code{simdat.c}}{Simulated responses of 500 examinees}
\item{\code{simalpha.c}}{Simulated attribute patterns of 500 examinees}
\item{\code{simQ}}{Simulated Q-matrix}
\item{\code{simcatprob.parm}}{Simulated success probabilities for each latent group in each item}
\item{\code{simdelta.parm}}{Simulated delta parameters for each item}
\item{\code{specifications}}{A list that contains all the specifications that were used in the \code{gen.itembank} function}
}
}
\usage{
sim180DINA
}
\description{
Simulated data, Q-matrix and item parameters for a 180-item bank with 5 attributes. Data generated using the \code{gen.itembank} function.
}
\keyword{datasets}
