% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim180GDINA.R
\docType{data}
\name{sim180GDINA}
\alias{sim180GDINA}
\title{Simulated data (180 items, G-DINA model)}
\format{
A list with components:
\describe{
\item{\code{simdat}}{Numeric matrix. Simulated responses of 500 examinees}
\item{\code{simQ}}{Simulated Q-matrix}
\item{\code{simalpha}}{Numeric matrix. Simulated attribute patterns of 500 examinees}
\item{\code{specifications}}{A list that contains all the specifications that were used in the \code{gen.itembank} function}
}
}
\usage{
sim180GDINA
}
\description{
Simulated data, Q-matrix and item parameters for a 180-item bank with 5 attributes. Data generated using the \code{gen.data} function.
}
\keyword{datasets}
